+++ components/core/log/src/commonMain/kotlin/com/flipperdevices/core/log/LogTagProvider.common.kt
@@ -4,7 +4,6 @@ package com.flipperdevices.core.log
 interface LogTagProvider {
     @Suppress("VariableNaming")
     val TAG: String
-    class Default(override val TAG: String) : LogTagProvider
 }
 
 expect inline fun LogTagProvider.error(logMessage: () -> String)
